Update ImageReader to reaad images without resizinf

+++ Toolkit.Foundation/IImageDescriptor.cs
@@ -4,7 +4,7 @@ public interface IImageDescriptor
 {
     object Image { get; }
 
-    int Width { get; }
+    double Width { get; }
 
-    int Height { get; }
+    double Height { get; }
 }
